requirements.txt
asgiref==3.5.2
Django==4.1.1
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.4
